+++ contracts/ElectoralSystems/ElectoralSystemInterface.sol
@@ -1,10 +1,10 @@
 pragma solidity 0.4.18;
 
-import "../OrganizationInterface.sol";
+import "../Proposals/Ballot/BallotInterface.sol";
 
 // @todo this is ugly, lets see if we can do it with libraries later
 interface ElectoralSystemInterface {
 
-    function winner(OrganizationInterface organization, uint id) public view returns (uint);
+    function winner(BallotInterface ballot) public view returns (uint);
 
 }