+++ contracts/OrganizationInterface.sol
@@ -1,7 +1,6 @@
 pragma solidity 0.4.18;
 
 import "./Managers/ProposalManagerInterface.sol";
-import "./Managers/VotingManagerInterface.sol";
 
 interface OrganizationInterface {
 
@@ -11,7 +10,6 @@ interface OrganizationInterface {
     function execute(uint id) external;
     function tally(uint id) external;
     function proposalManager() external view returns (ProposalManagerInterface);
-    function votingManager() external view returns (VotingManagerInterface);
     function quorumReached(uint id) public view returns (bool);
     function winningOption(uint id) public view returns (uint);